+++ lib/src/apps/apps_header.dart
@@ -12,7 +12,7 @@ part of lightning_ctrl;
 class AppsHeader extends LComponent {
 
   final Element element = new Element.header()
-    ..classes.add(LMargin.C_AROUND__SMALL)
+    ..classes.addAll([LMargin.C_AROUND__SMALL, LGrid.C_GRID, LGrid.C_WRAP])
     ..id = "a-header";
 
   /**
@@ -20,9 +20,53 @@ class AppsHeader extends LComponent {
    */
   void set(AppsCtrl apps) {
     element.children.clear();
-    if (apps.imageSrc != null) {
+
+    // Logo
+    DivElement left = new DivElement()
+      ..classes.add(LGrid.C_COL);
+    element.append(left);
+    if (apps.icon != null) {
+      left.append(apps.icon.element);
+    }
+    else if (apps.imageSrc != null) {
       LImage img = new LImage.srcMedium(apps.imageSrc, apps.label, circle: false);
-      element.append(img.element);
+      left.append(img.element);
+    }
+
+    // Title
+    DivElement center = new DivElement()
+      ..classes.add(LGrid.C_COL__PADDED);
+    element.append(center);
+    HeadingElement h1 = new HeadingElement.h1()
+      ..classes.add(LText.C_TEXT_HEADING__MEDIUM)
+      ..text = apps.label;
+    center.append(h1);
+    document.title = apps.label;
+    if (apps.labelSub != null) {
+      DivElement sub = new DivElement()
+        ..classes.add(LText.C_TEXT_HEADING__SMALL)
+        ..text = apps.labelSub;
+      center.append(sub);
+    }
+
+    // Info
+    if (apps.info != null && apps.info.isNotEmpty) {
+      DivElement right = new DivElement()
+        ..classes.add(LGrid.C_COL);
+      element.append(right);
+      List<String> parts = apps.info.split("\n");
+      for (String part in parts) {
+        if (right.children.isNotEmpty) {
+          right.append(new Element.br());
+        }
+        right.append(new SpanElement()
+          ..text = part);
+      }
+    }
+
+    /// Additional Header Element
+    if (apps.additionalHeader != null) {
+      element.append(apps.additionalHeader);
     }
   } // set